feat: remove useAppTheme dependent code & rename files with _
