+++ packages/design-system/src/components/button/index.ts
@@ -1,2 +1,2 @@
-export * from "./button";
-export { default } from "./button";
+export * from "./_button";
+export { default } from "./_button";
